Fix key import via URL when Internet Explorer engine is not available
Description
===========

I had an error running the role to setup auth keys with github key urls.
After tweeking the code to print the exception message instead of the
generic error I received this error:

```
The response content cannot be parsed because the Internet Explorer engine is
not available, or Internet Explorer's first-launch configuration is not complete.
Specify the UseBasicParsing parameter and try again.
```

I tried it and it works. So this patch just makes sure to use
`-UseBasicParsing` flag in the `invoke-webrequest` calls.